requirements.txt
# Core FastAPI and web framework
fastapi==0.115.12
uvicorn==0.34.3
starlette==0.46.2
python-multipart==0.0.20

# Database
sqlalchemy==2.0.41
databases==0.9.0
aiosqlite==0.21.0
greenlet==3.2.2

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# Validation (Pydantic v2)
pydantic==2.11.5
pydantic_core==2.33.2
annotated-types==0.7.0

# Core dependencies
anyio==4.9.0
sniffio==1.3.1
idna==3.10
click==8.2.1
colorama==0.4.6
typing_extensions==4.14.0
typing-inspection==0.4.1

# Template engine
jinja2==3.1.6
MarkupSafe==3.0.2

# Configuration
python-dotenv==1.1.0
PyYAML==6.0.2

# Server features
httptools==0.6.4
watchfiles==1.0.5
websockets==15.0.1
h11==0.16.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
requests==2.32.3